Add files via upload
Lots more additions. All that needs to be done now is SQL statements to database and sorting of arrays for each filter option. Mobile's formatting is all set but I would like to change the desktop view to not use checkboxes as that would suggest multiple filters and that's not how I wrote the code. We need to have a discussion on multiple checkboxes in person though. If we were to allow multiple filters we would have to set arbitrary values for those thresholds and we lose lots of modularity. We lose the ability to say this project could be used by any company with any product though we don't have a strong basis for that as it is.