Product adding/deleting should work now
Just need to change sql statements if your DB is set up differently.
Adding a product adds it to database and uploads a picture to /images.
Deleting a products deletes the row from the databse and deletes the
picture from the database. Products that use the same picture will cause
problems when one is deleted.